{"schema":"decodex/commit/1","summary":"Split execution architecture recovery","authority":"XY-1115"}
